fix run job

+++ subscriptions/src/db/postgres.py
@@ -1,6 +1,6 @@
+import typing
 
-from sqlalchemy.ext.asyncio import AsyncEngine
-from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession
 from sqlalchemy.orm import sessionmaker
 
 from models import Base
@@ -9,16 +9,25 @@ engine: AsyncEngine | None = None
 async_session: sessionmaker | None = None
 
 
-async def get_session() -> AsyncSession:
+async def get_session() -> typing.AsyncGenerator[AsyncSession, None]:
+    if not async_session:
+        raise RuntimeError("Database is not initialized")
+
     async with async_session() as session:
         yield session
 
 
 async def create_database() -> None:
+    if not engine:
+        raise RuntimeError("Database is not initialized")
+
     async with engine.begin() as conn:
         await conn.run_sync(Base.metadata.create_all)
 
 
 async def purge_database() -> None:
+    if not engine:
+        raise RuntimeError("Database is not initialized")
+
     async with engine.begin() as conn:
         await conn.run_sync(Base.metadata.drop_all)
